+++ examples/demo/delta_tweens.rs
@@ -11,7 +11,10 @@ fn secs(secs: f32) -> Duration {
 
 fn main() {
     App::new()
-        .add_plugins((DefaultPlugins, DefaultTweenPlugins))
+        .add_plugins((
+            DefaultPlugins,
+            DefaultTweenPluginsOnDefaultTime::default(),
+        ))
         .add_systems(Startup, (setup, spawn_circle_with_tweens))
         .run();
 }